+++ EmptyClass.cpp
@@ -1,0 +1,12 @@
+/*
+* EmptyClass.cpp
+*
+*  Created on: <Date>
+*      Author: <Author>
+*/
+
+#include "EmptyClass.h"
+
+EmptyClass::EmptyClass() {
+
+}