+++ EmptyClass.h
@@ -1,0 +1,23 @@
+ /*
+ * EmptyClass.h
+ *
+ *  Created on: <Date>
+ *      Author: <Author>
+ */
+
+#ifndef EMPTYCLASS_H_
+#define EMPTYCLASS_H_
+
+#include "gApp.h"
+
+
+class EmptyClass {
+	public:
+	EmptyClass();
+
+	private:
+    int x;
+};
+
+
+#endif /* EMPTYCLASS_H_ */