+++ src/extension.ts
@@ -4,8 +4,9 @@ import * as fs from 'fs-extra';
 import * as path from 'path';
 import AdmZip from 'adm-zip';
 import axios from 'axios';
+import json5 from 'json5'
 import { execSync } from 'child_process';
-
+import { rimraf } from 'rimraf';
 
 
 const currentDirectory = process.cwd();
@@ -25,49 +26,57 @@ const glistClangUrl = "https://github.com/javertus/glistzbin-win64-vscode/releas
 const glistCmakeUrl = "https://github.com/javertus/glistzbin-win64-vscode/releases/download/Dependencies/CMake.zip";
 const ninjaUrl = "https://github.com/ninja-build/ninja/releases/download/v1.12.1/ninja-win.zip";
 
-let jsonParsedData: any;
+let extensionJsonData: any;
 let extensionPath: string;
 let extensionDataFilePath: string;
 
 export function activate(context: vscode.ExtensionContext) {
-	vscode.commands.registerCommand('glist-engine-worker-extension.create-glistapp', async () => {
-		await CreateGlistApp();
+	vscode.commands.registerCommand('glist-extension.create-glistapp', async () => {
+		await CreateNewProject();
 	});
-	vscode.commands.registerCommand('glist-engine-worker-extension.workspace-update', async () => {
+	vscode.commands.registerCommand('glist-extension.workspace-update', async () => {
 		await UpdateWorkspace();
 	});
-	vscode.commands.registerCommand('glist-engine-worker-extension.openWorkspace', async () => {
+	vscode.commands.registerCommand('glist-extension.openWorkspace', async () => {
 		await LaunchWorkspace();
 	});
-	vscode.commands.registerCommand('glist-engine-worker-extension.reset', async () => {
+	vscode.commands.registerCommand('glist-extension.reset', async () => {
 		ResetExtension();
 	});
-	vscode.commands.registerCommand('glist-engine-worker-extension.installglistengine', async () => {
-		InstallGlistEngine();
+	vscode.commands.registerCommand('glist-extension.installglistengine', async () => {
+		extensionJsonData.installGlistEngine = true;
+		SaveExtensionJson()
+		if (await updateVSCodeSettings()) return;
+		extensionJsonData.installGlistEngine = false;
+		SaveExtensionJson()
+		await InstallGlistEngine();
 	});
-
+	vscode.commands.registerCommand('glist-extension.deleteproject', async () => {
+		await DeleteProject();
+	});
+	vscode.commands.registerCommand('glist-extension.addcanvastoproject', async () => {
+		await AddClassToProject(path.join(extensionPath, "GlistApp", "src"), "gCanvas");
+	});
+	vscode.commands.registerCommand('glist-extension.addclasstoproject', async () => {
+		await AddClassToProject(path.join(extensionPath), "EmptyClass");
+	});
 	extensionPath = context.extensionPath;
 	extensionDataFilePath = path.join(extensionPath, 'ExtensionData.json');
 	FirstRunWorker();
 }
 
-async function CreateGlistApp(projectName: any = undefined) {
+async function CreateNewProject(projectName: any = undefined) {
 	if (projectName == undefined) {
 		projectName = await vscode.window.showInputBox({
-			placeHolder: "Enter the name of new GlistApp"
+			placeHolder: "Enter the name of new Project"
 		});
 	}
-
-	if (!projectName) {
-		vscode.window.showErrorMessage("No input provided.");
-		return;
+	if (checkInput(projectName)) return;
+	projectName = projectName + "";
+	if (!checkPath(path.join(glistappsPath, projectName), "A project named " + projectName + " already exist. Opening already existing project...", false)) {
+		fs.cpSync(path.join(extensionPath, 'GlistApp'), path.join(glistappsPath, projectName), { recursive: true });
+		vscode.window.showInformationMessage('Created new Project.');
 	}
-	if (fs.existsSync(path.join(glistappsPath, projectName))) {
-		vscode.window.showErrorMessage("A folder named " + projectName + " already exist.");
-		return;
-	}
-
-	fs.cpSync(path.join(extensionPath, 'GlistApp'), path.join(glistappsPath, projectName), { recursive: true });
 
 	await AddNewProjectToWorkspace(projectName);
 
@@ -85,6 +94,13 @@ async function CreateGlistApp(projectName: any = undefined) {
 
 async function AddNewProjectToWorkspace(projectName: string) {
 	// Read the JSON file
+	if (!fs.existsSync(workspaceFilePath)) {
+		extensionJsonData.firstRun = false;
+		extensionJsonData.secondRun = true;
+		SaveExtensionJson()
+		UpdateWorkspace();
+		return;
+	}
 	const jsonString = fs.readFileSync(workspaceFilePath, 'utf-8');
 	const jsonData = JSON.parse(jsonString);
 
@@ -103,10 +119,33 @@ async function AddNewProjectToWorkspace(projectName: string) {
 
 	// Convert the updated data back to JSON format and write it to the file
 	fs.writeFileSync(workspaceFilePath, JSON.stringify(jsonData, null, 2));
-	vscode.window.showInformationMessage('Created new GlistApp.');
 
 	//VS Code will restart if another workspace is active.
 	await vscode.commands.executeCommand('vscode.openFolder', vscode.Uri.file(workspaceFilePath), false);
+}
+
+async function DeleteProjectFromWorkspace(projectName: string): Promise<Boolean> {
+	if (!fs.existsSync(workspaceFilePath)) {
+		vscode.window.showWarningMessage('Workspace file does not exist.');
+		return false;
+	}
+
+	const jsonString = fs.readFileSync(workspaceFilePath, 'utf-8');
+	const jsonData = JSON.parse(jsonString);
+	// Check if the project exists in the workspace folders
+	const projectIndex = jsonData.folders.findIndex((folder: { path: string }) =>
+		folder.path.includes(glistappsPath) && folder.path.endsWith(`${projectName}`)
+	);
+	if (projectIndex === -1) {
+		return true;
+	}
+	jsonData.folders = jsonData.folders.filter((folder: { path: string }) =>
+		!(folder.path.includes(glistappsPath) && folder.path.endsWith(`${projectName}`))
+	);
+
+	await fs.writeFile(workspaceFilePath, JSON.stringify(jsonData, null, 2));
+	vscode.window.showInformationMessage(`Deleted project '${projectName}' from workspace.`);
+	return true;
 }
 
 async function UpdateWorkspace() {
@@ -147,10 +186,11 @@ async function LaunchWorkspace() {
 }
 
 function ResetExtension() {
-	jsonParsedData.firstRun = true;
-	jsonParsedData.secondRun = true;
-	fs.writeFileSync(extensionDataFilePath, JSON.stringify(jsonParsedData, null, 2));
+	extensionJsonData.firstRun = true;
+	extensionJsonData.secondRun = true;
+	SaveExtensionJson()
 	fs.rmSync(path.join(extensionPath, 'GlistApp'), { recursive: true, force: true });
+	vscode.commands.executeCommand('workbench.action.reloadWindow');
 }
 
 async function InstallGlistEngine() {
@@ -169,14 +209,14 @@ async function InstallGlistEngine() {
 			progress.report({ increment: 0 });
 			await createDirectories();
 			progress.report({ increment: 20 });
-			await installGlistEngine();
+			await InstallEngine();
 			progress.report({ increment: 20 });
-			await installCmake();
+			await InstallCmake();
 			progress.report({ increment: 20 });
-			AddPathToSystemEnvironment(path.join(glistZbinPath, "CMake"));
-			await installClang();
+			AddPathToSystemEnvironment(path.join(glistZbinPath, "CMake", "bin"));
+			await InstallClang();
 			progress.report({ increment: 20 });
-			await createEmptyGlistApp();
+			await createEmptyProject();
 			progress.report({ increment: 20 });
 			vscode.window.showInformationMessage("Glist Engine Installed Successfully.");
 		});
@@ -195,14 +235,13 @@ async function DownloadFile(url: string, dest: string) {
 	fs.writeFileSync(dest, response.data);
 }
 
-
 function ExtractArchive(zipPath: string, dest: string, message: string) {
 	const zip = new AdmZip(zipPath);
 	zip.extractAllTo(dest, true);
 	vscode.window.showInformationMessage(message);
 }
 
-async function installGlistEngine() {
+async function InstallEngine() {
 	vscode.window.showInformationMessage("Installing Engine ~8MB");
 
 	const zipFilePath = path.join(tempPath, 'GlistEngine.zip');
@@ -212,7 +251,7 @@ async function installGlistEngine() {
 	await fs.rename(path.join(glistPath, 'GlistEngine-main'), path.join(glistPath, 'GlistEngine'));
 }
 
-async function installCmake() {
+async function InstallCmake() {
 	vscode.window.showInformationMessage("Installing Cmake ~35MB");
 
 	const zipFilePath = path.join(tempPath, 'CMake.zip');
@@ -221,7 +260,7 @@ async function installCmake() {
 	ExtractArchive(zipFilePath, glistZbinPath, "CMake Binaries Installed.");
 }
 
-async function installClang() {
+async function InstallClang() {
 	vscode.window.showInformationMessage("Installing Clang Binaries ~400MB");
 
 	const zipFilePath = path.join(tempPath, 'clang64.zip');
@@ -231,60 +270,262 @@ async function installClang() {
 }
 
 function AddPathToSystemEnvironment(newPath: string) {
-	const currentPath = process.env.PATH || '';
+	const command = `powershell.exe -Command "[Environment]::GetEnvironmentVariable('Path', 'User')"`;
+	//Removing the last index because it's space 
+	let currentPath = execSync(command, { encoding: 'utf-8' }).replace(/\n/g, '').slice(0, -1);
+	if (currentPath.endsWith(';')) currentPath = currentPath.slice(0, -1);
 	if (!currentPath.includes(newPath)) {
 		const psCommand = `powershell -Command "[Environment]::SetEnvironmentVariable('Path', '${currentPath};${newPath}', [EnvironmentVariableTarget]::User)"`;
 		execSync(psCommand);
 	}
 }
-async function createEmptyGlistApp() {
-	vscode.window.showInformationMessage("Creating Empty GlistApp");
-	CreateGlistApp("GlistApp");
+
+async function updateVSCodeSettings(): Promise<boolean> {
+	const vscodeSettingsPath = path.join(
+		process.env.HOME || process.env.USERPROFILE || '',
+		'AppData/Roaming/Code/User/settings.json' // This path might differ based on OS and VS Code distribution
+	);
+	let settings: any;
+	const newSettings = {
+		"extensions.ignoreRecommendations": true,
+		"cmake.options.statusBarVisibility": "visible",
+		"cmake.showOptionsMovedNotification": false,
+		"cmake.configureOnOpen": true,
+		"security.workspace.trust.enabled": false,
+		"security.workspace.trust.banner": "never",
+		"security.workspace.trust.untrustedFiles": "open",
+		"security.workspace.trust.startupPrompt": "never",
+	};
+	try {
+		if (fs.existsSync(vscodeSettingsPath)) {
+			const fileContent = await fs.readFile(vscodeSettingsPath, 'utf-8');
+			settings = json5.parse(fileContent);
+		}
+		else {
+			await fs.writeFile(vscodeSettingsPath, JSON.stringify(newSettings, null, 2));
+			vscode.commands.executeCommand('workbench.action.reloadWindow');
+			return true;
+		}
+
+		let isChanged = false;
+
+		// Add or update settings
+		for (const [key, value] of Object.entries(newSettings)) {
+			if (settings[key] !== value) {
+				settings[key] = value;
+				isChanged = true;
+			}
+		}
+		if (isChanged) {
+			await fs.writeFile(vscodeSettingsPath, JSON.stringify(settings, null, 2));
+			vscode.commands.executeCommand('workbench.action.reloadWindow');
+		}
+		return isChanged;
+	} catch (err) {
+		console.error('Error while updating VS Code settings:', err);
+		return true;
+	}
+}
+
+function ReplaceText(inputFilePath: string, searchText: string, replaceText: string) {
+	let data = fs.readFileSync(inputFilePath, 'utf8')
+	const result = data.replace(new RegExp(searchText, 'g'), replaceText);
+	fs.writeFileSync(inputFilePath, result, 'utf8');
+}
+
+function AddFileToCMakeLists(projectPath: string, newFileName: string) {
+	const cmakeListsPath = path.join(projectPath, 'CMakeLists.txt');
+
+	let data = fs.readFileSync(cmakeListsPath, 'utf8');
+
+	// Add new file to GlistApp_SOURCES
+	const sourcesPattern = /set\(GlistApp_SOURCES\s*\n([^)]*)\)/;
+	const sourcesMatch = data.match(sourcesPattern);
+	if (sourcesMatch && sourcesMatch[1]) {
+		const sources = sourcesMatch[1];
+		const newSources = sources.trim() + "\n\t${APP_DIR}" + `/src/${newFileName}.cpp`;
+		data = data.replace(sourcesPattern, `set(GlistApp_SOURCES\n${newSources}\n)`);
+	}
+
+	// Add new file to GlistApp_HEADERS
+	const headersPattern = /set\(GlistApp_HEADERS\s*\n([^)]*)\)/;
+	const headersMatch = data.match(headersPattern);
+	if (headersMatch && headersMatch[1]) {
+		const headers = headersMatch[1];
+		const newHeaders = headers.trim() + "\n\t${APP_DIR}" + `/src/${newFileName}.h`;
+		data = data.replace(headersPattern, `set(GlistApp_HEADERS\n${newHeaders}\n)`);
+	}
+
+	fs.writeFileSync(cmakeListsPath, data, 'utf8');
+}
+
+
+async function createEmptyProject() {
+	vscode.window.showInformationMessage("Creating Empty Project");
+	CreateNewProject("GlistApp");
 }
 
 async function createDirectories() {
+	await fs.ensureDir(glistappsPath);
 	await fs.ensureDir(glistpluginsPath);
 	await fs.remove(tempPath);
 	await fs.ensureDir(tempPath);
 }
 
+async function CloseNonExistentFileTabs() {
+	const visibleWindows = vscode.window.tabGroups.activeTabGroup.tabs;
+	try {
+		for (const visibleWindow of visibleWindows) {
+			const filePath = JSON.parse(JSON.stringify(visibleWindow.input, null, 2));
+			if (fs.existsSync(filePath.uri.fsPath)) continue;
+			await vscode.window.showTextDocument(filePath, { preview: false });
+		}
+	}
+	catch (err) {
+		console.log(err);
+		await vscode.commands.executeCommand('workbench.action.closeActiveEditor');
+		CloseNonExistentFileTabs();
+	}
+}
+
+async function DeleteFolder(folderPath: string): Promise<void> {
+	await rimraf(folderPath);
+	console.log(`Folder ${folderPath} deleted successfully.`);
+}
+
+async function DeleteProject() {
+	let projectName = await vscode.window.showInputBox({
+		placeHolder: "Enter the name of Project you want to delete"
+	});
+	if (checkInput(projectName)) return;
+	projectName = projectName + "";
+	if (checkPath(path.join(glistappsPath, projectName), "A project named " + projectName + " does not exist!")) return;
+
+	let decision = await vscode.window.showInputBox({
+		placeHolder: 'Are you sure about deleting this project? Type "yes" to continue.'
+	});
+	if (!(decision?.toLowerCase() == "yes")) {
+		vscode.window.showErrorMessage("Deleting Project Cancelled!");
+		return;
+	}
+	await DeleteProjectFromWorkspace(projectName)
+	DeleteFolder(path.join(glistappsPath, projectName));
+	extensionJsonData.deleteFolder = path.join(glistappsPath, projectName);
+	fs.writeFile(extensionDataFilePath, JSON.stringify(extensionJsonData, null, 2));
+	vscode.commands.executeCommand('workbench.action.reloadWindow');
+}
+
+function checkInput(name: any, message: string = "No input provided."): Boolean {
+	if (!name) {
+		vscode.window.showErrorMessage(message);
+		return true;
+	}
+	return false;
+}
+
+function checkPath(inputPath: any, message: any = undefined, errorMessageIfNotExist = true) {
+	if (!fs.existsSync(inputPath) && errorMessageIfNotExist) {
+		vscode.window.showErrorMessage(message);
+		return true;
+	}
+	else if (fs.existsSync(inputPath) && errorMessageIfNotExist == false) {
+		vscode.window.showErrorMessage(message);
+		return true;
+	}
+	return false;
+}
+
+async function AddClassToProject(baseFilePath: string, fileBaseName: string) {
+	let projectName = await vscode.window.showInputBox({
+		placeHolder: "Enter the name of project you want to create new class"
+	});
+	if (!projectName) {
+		vscode.window.showErrorMessage("No input provided.");
+		return;
+	}
+	if (checkPath(path.join(glistappsPath, projectName), "A project named " + projectName + " does not exist!")) return;
+
+	let className = await vscode.window.showInputBox({
+		placeHolder: "Enter the name of canvas you want to create"
+	});
+	if (checkInput(className)) return;
+	className = className + "";
+	if (checkPath(path.join(glistappsPath, projectName, "src", className + ".h"), "A class named " + className + " already exist!", false)) return;
+	if (checkPath(path.join(glistappsPath, projectName, "src", className + ".cpp"), "A class named " + className + " already exist!", false)) return;
+
+	fs.copyFileSync(path.join(baseFilePath, fileBaseName + ".h"), path.join(glistappsPath, projectName, "src", className + ".h"));
+	fs.copyFileSync(path.join(baseFilePath, fileBaseName + ".cpp"), path.join(glistappsPath, projectName, "src", className + ".cpp"));
+	ReplaceText(path.join(glistappsPath, projectName, "src", className + ".h"), fileBaseName, className);
+	ReplaceText(path.join(glistappsPath, projectName, "src", className + ".cpp"), fileBaseName, className);
+	ReplaceText(path.join(glistappsPath, projectName, "src", className + ".h"), fileBaseName.toUpperCase() + "_H_", className.toUpperCase() + "_H_");
+	AddFileToCMakeLists(path.join(glistappsPath, projectName), className);
+
+	const filesToOpen = [
+		path.join(glistappsPath, projectName, 'src', className + ".h"),
+		path.join(glistappsPath, projectName, 'src', className + ".cpp")
+	];
+
+	filesToOpen.forEach(async file => {
+		const uri = vscode.Uri.file(file);
+		const document = await vscode.workspace.openTextDocument(uri);
+		await vscode.window.showTextDocument(document, { preview: false });
+	});
+}
+
+function SaveExtensionJson() {
+	fs.writeFileSync(extensionDataFilePath, JSON.stringify(extensionJsonData, null, 2));
+}
 
 async function FirstRunWorker() {
 	CheckJsonFile();
-	if (jsonParsedData.firstRun) {
+	if (extensionJsonData.deleteFolder) {
+		await DeleteFolder(extensionJsonData.deleteFolder);
+		extensionJsonData.deleteFolder = undefined;
+		SaveExtensionJson()
+		vscode.window.showInformationMessage("Project Deleted.");
+	}
+	await CloseNonExistentFileTabs();
+	if (extensionJsonData.installGlistEngine) {
+		extensionJsonData.installGlistEngine = false;
+		SaveExtensionJson()
+		await InstallGlistEngine();
+	}
+	if (!fs.existsSync(path.join(extensionPath, "GlistApp"))) {
+		await InstallGlistAppTemplate();
+	}
+	if (extensionJsonData.firstRun) {
 		vscode.window.withProgress({
 			location: vscode.ProgressLocation.Notification,
 			cancellable: false,
 			title: 'Installing required extensions for Glist Engine. Please wait...'
 		}, async (progress) => {
 			progress.report({ increment: 0 });
-			const zipFilePath = path.join(extensionPath, "GlistApp.zip");
-			await DownloadFile(glistAppUrl, zipFilePath);
-			ExtractArchive(zipFilePath, extensionPath, "");
-			await fs.rename(path.join(extensionPath, 'GlistApp-vscode-main'), path.join(extensionPath, 'GlistApp'));
-			await fs.remove(zipFilePath);
 			await InstallExtensions(progress);
 			await CreateWorkspace();
 		});
 	}
-	else if (jsonParsedData.secondRun) {
+	else if (extensionJsonData.secondRun) {
 		await OpenFiles();
 	}
 }
 
 function CheckJsonFile() {
-	// Check if the file exists
 	if (!fs.existsSync(extensionDataFilePath)) {
-		// File does not exist, create it with initial data
-		let initialData = { firstRun: true, secondRun: true };
+		let initialData = { firstRun: true, secondRun: true, installGlistEngine: false, deleteFolder: undefined };
 		fs.writeFileSync(extensionDataFilePath, JSON.stringify(initialData, null, 2));
 	}
-	//Read data from file
 	let data = fs.readFileSync(extensionDataFilePath, 'utf8');
-	jsonParsedData = JSON.parse(data);
+	extensionJsonData = JSON.parse(data);
 }
 
-// Installs the required extensions
+async function InstallGlistAppTemplate() {
+	const zipFilePath = path.join(extensionPath, "GlistApp.zip");
+	await DownloadFile(glistAppUrl, zipFilePath);
+	ExtractArchive(zipFilePath, extensionPath, "");
+	await fs.rename(path.join(extensionPath, 'GlistApp-vscode-main'), path.join(extensionPath, 'GlistApp'));
+	await fs.remove(zipFilePath);
+}
+
 async function InstallExtensions(progress: any) {
 	try {
 		// Required extension names
@@ -314,25 +555,28 @@ async function InstallExtensions(progress: any) {
 	}
 }
 
-// Create and open a workspace for glist engine
 async function CreateWorkspace() {
 	try {
 		// Do not create workspace and stop setup process if glistapps does not exist (Also Meaning Glist Engine is not installed.)
-		if(!fs.existsSync(glistappsPath)) {
-			jsonParsedData.firstRun = false;
-			jsonParsedData.secondRun = false;
-			fs.writeFileSync(extensionDataFilePath, JSON.stringify(jsonParsedData, null, 2));
+		if (!fs.existsSync(glistappsPath)) {
+			extensionJsonData.firstRun = false;
+			extensionJsonData.secondRun = false;
+			SaveExtensionJson()
 			return;
 		}
+		await updateVSCodeSettings();
 		// Install ninja if does not exist
 		fs.ensureDirSync(path.join(glistZbinPath, "CMake"))
-		const ninjaPath = path.join(glistZbinPath, "CMake", "ninja.zip");
-		if(!fs.existsSync(ninjaPath)) {
-			vscode.window.showInformationMessage("yok");
+		const ninjaPath = path.join(glistZbinPath, "CMake", "bin", "ninja.zip");
+		if (!fs.existsSync(path.join(glistZbinPath, "CMake", "bin", "ninja.exe"))) {
+			vscode.window.showInformationMessage("Ninja not found. Installing...");
 			await DownloadFile(ninjaUrl, ninjaPath);
-			ExtractArchive(ninjaPath, path.join(glistZbinPath, "CMake"), "");
+			ExtractArchive(ninjaPath, path.join(glistZbinPath, "CMake", "bin"), "");
 			await fs.remove(ninjaPath);
 		}
+
+		// Add cmake to system path
+		AddPathToSystemEnvironment(path.join(glistZbinPath, "CMake", "bin"));
 
 		let workspaceFolders = [];
 		getSubfolders(glistappsPath).map(folder => {
@@ -351,8 +595,8 @@ async function CreateWorkspace() {
 
 		fs.writeFileSync(workspaceFilePath, JSON.stringify(workspaceContent, null, 2));
 		vscode.window.showInformationMessage('Workspace configured.');
-		jsonParsedData.firstRun = false;
-		fs.writeFileSync(extensionDataFilePath, JSON.stringify(jsonParsedData, null, 2));
+		extensionJsonData.firstRun = false;
+		SaveExtensionJson()
 
 		// Opens the new workspace. Setup cannot continue after here because vscode restarts. For resuming setup, there is a secondary setup run.
 		await vscode.commands.executeCommand('vscode.openFolder', vscode.Uri.file(workspaceFilePath), false);
@@ -382,8 +626,8 @@ async function OpenFiles() {
 		await vscode.window.showTextDocument(document, { preview: false });
 	});
 
-	jsonParsedData.secondRun = false;
-	fs.writeFileSync(extensionDataFilePath, JSON.stringify(jsonParsedData, null, 2));
+	extensionJsonData.secondRun = false;
+	SaveExtensionJson()
 }
 
 function getSubfolders(directory: string): string[] {