+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+# 1.2.0
+
+## New
+- Creating new canvas or empty class to given project
+- Closing non-existent file tabs
+- Deleting projects
+- Various bug fixes
+
 # 1.1.0
 
 ## New